pyproject.toml
[project]
name = "ipum-lab-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=3.0.1",
    "pre-commit>=4.5.1",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
]
